Title: 바이러스 검사, Time: 6ms, Memory: 0MB, Status: Wrong Answer - Codetree

+++ 250410/바이러스 검사/virus-detector.cpp
@@ -29,7 +29,7 @@ int main() {
     //solve
     ll sm = 0;
     for(int i = 0; i < n; i++) {
-        int k = ceil(((float)customePerStore[i] - a) / b);
+        ll k = ceil(((float)customePerStore[i] - a) / b);
         sm += (k + 1);
     }
     cout << sm;
